Refactor output and progress into separate packages in run
