+++ bundle/run/output/run.go
@@ -1,0 +1,5 @@
+package output
+
+type RunOutput interface {
+	String() (string, error)
+}
