+++ bundle/run/output/task.go
@@ -1,4 +1,4 @@
-package run
+package output
 
 import (
 	"encoding/json"
