+++ bundle/run/output/task_test.go
@@ -1,4 +1,4 @@
-package run
+package output
 
 import (
 	"fmt"
