+++ bundle/run/progress/pipeline.go
@@ -1,4 +1,4 @@
-package pipeline
+package progress
 
 import (
 	"context"
