+++ bundle/run/progress/pipeline_events.go
@@ -1,4 +1,4 @@
-package pipeline
+package progress
 
 import "fmt"
