+++ bundle/run/runner.go
@@ -6,18 +6,13 @@ import (
 	"strings"
 
 	"github.com/databricks/bricks/bundle"
+	"github.com/databricks/bricks/bundle/run/output"
 )
 
-// TODO: refactor this package into service specific implementations. Its
-// getting bloated. (https://github.com/databricks/bricks/issues/282)
 type key string
 
 func (k key) Key() string {
 	return string(k)
-}
-
-type RunOutput interface {
-	String() (string, error)
 }
 
 // Runner defines the interface for a runnable resource (or workload).
@@ -27,7 +22,7 @@ type Runner interface {
 	Key() string
 
 	// Run the underlying worklow.
-	Run(ctx context.Context, opts *Options) (RunOutput, error)
+	Run(ctx context.Context, opts *Options) (output.RunOutput, error)
 }
 
 // Find locates a runner matching the specified argument.
